Booking functionality web with socket

+++ app/WEB/forms/journey.py
@@ -7,8 +7,8 @@ from wtforms.widgets import HiddenInput, Select
 class CreateJourneyForm(FlaskForm):
     from_ = StringField("From", validators=[DataRequired()])
     to = StringField("To", validators=[DataRequired()])
-    distance = IntegerField("Distance (Kilometers)")
-    duration = IntegerField("Duration (Hours)")
+    distance = FloatField("Distance (Kilometers)")
+    duration = FloatField("Duration (Hours)")
     submit = SubmitField('Save')
 
     def __init__(self, *args, **kwargs):
@@ -19,8 +19,8 @@ class UpdateJourneyForm(FlaskForm):
 	id = IntegerField(validators=[DataRequired()], widget=HiddenInput())
 	from_ = StringField("From", validators=[DataRequired()])
 	to = StringField("To", validators=[DataRequired()])
-	distance = IntegerField("Distance (Kilometers)")
-	duration = IntegerField("Duration (Hours)")
+	distance = FloatField("Distance (Kilometers)")
+	duration = FloatField("Duration (Hours)")
 	submit = SubmitField('Save')
 
 	def __init__(self, *args, **kwargs):
